+++ lib/sbmt/kafka_consumer/serialization/json_deserializer.rb
@@ -8,8 +8,7 @@ module Sbmt
           # nil payload can be present for example for tombstone messages
           message.raw_payload.nil? ? nil : ::JSON.parse(message.raw_payload)
         rescue JSON::ParserError => e
-          ::Sbmt::KafkaConsumer.logger.error("decoding error: #{e.message}")
-          return if skip_decoding_error
+          raise Sbmt::KafkaConsumer::SkipUndeserializableMessage, "cannot decode message: #{e.message}, payload: #{message.raw_payload}" if skip_decoding_error
 
           raise
         end